src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-cp-vcenter>=2.2.0,<2.3.0
cloudshell-cp-core>=1.1.0,<1.2.0
